+++ deltafi-core/src/main/java/org/deltafi/core/schedulers/trigger/ConfigurableFixedDelayTrigger.java
@@ -46,6 +46,6 @@ public class ConfigurableFixedDelayTrigger implements Trigger {
             return triggerContext.getClock().instant().plusMillis(this.initialDelay);
         }
 
-        return lastCompletion.plusMillis(durationFunction.apply(deltaFiPropertiesService.getDeltaFiProperties()).toMillis());
+        return lastExecution.plusMillis(durationFunction.apply(deltaFiPropertiesService.getDeltaFiProperties()).toMillis());
     }
 }